+++ src/core/core.module.ts
@@ -1,7 +1,7 @@
 import { Global, Logger, Module } from "@nestjs/common";
 import { APP_FILTER, APP_INTERCEPTOR } from "@nestjs/core";
 import { AppConfigService } from "@src/core/configs/app-config.service";
-import { HttpExceptionFilter } from "@src/core/filter/http-exception.filter";
+import { AllExceptionsFilter } from "@src/core/filter/all-exceptions.filter";
 import { ResponseInterceptor } from "@src/core/interceptors/response.interceptor";
 import { createWinstonLogger } from "@src/core/logger/winston.config";
 import { WinstonLogger } from "@src/core/logger/winston.logger";
@@ -22,7 +22,7 @@ import { BootstrapService } from "@src/core/services/bootstrap.service";
     },
     {
       provide: APP_FILTER,
-      useClass: HttpExceptionFilter,
+      useClass: AllExceptionsFilter,
     },
     {
       provide: APP_INTERCEPTOR,
